* (***ADDB***) Implement auto-update for incompatible outdated versions (auto-enabled by default). * (***ADDB***) Implement always auto-update (tries to keep you on the latest version) (auto-disabled by default). * (***ADDB***) Remove UMM Logger patch (finally no more issue for Doorstop?).

+++ ToyBox/Classes/Infrastructure/Update/AutoUpdater.cs
@@ -1,5 +1,4 @@
-﻿using Kingmaker.GameInfo;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -13,7 +12,7 @@ using UnityModManagerNet;
 
 namespace ToyBox {
     public static class AutoUpdater {
-        public static bool Update(UnityModManager.ModEntry.ModLogger logger, string repoLink, string repositoryJsonLink, string releaseName) {
+        public static bool Update(UnityModManager.ModEntry.ModLogger logger, string repoLink, string repositoryJsonLink, string releaseName, string curVersion) {
             try {
                 using var web = new WebClient();
                 var curDir = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
@@ -32,26 +31,30 @@ namespace ToyBox {
                 var raw = web.DownloadString(repositoryJsonLink);
                 var result = JsonConvert.DeserializeAnonymousType(raw, definition);
                 string version = result.Releases[0].Version;
-                string downloadUrl = $"{repoLink}/releases/download/v{version}/{releaseName}-{version}.zip";
+                if (new Version(VersionChecker.GetNumifiedVersion(logger, version)) > new Version(VersionChecker.GetNumifiedVersion(logger, curVersion))) {
+                    string downloadUrl = $"{repoLink}/releases/download/v{version}/{releaseName}-{version}.zip";
+                    logger.Log($"Downloading: {downloadUrl}");
+                    web.DownloadFile(downloadUrl, file.FullName);
+                    using var zipFile = ZipFile.OpenRead(file.FullName);
+                    foreach (ZipArchiveEntry entry in zipFile.Entries) {
+                        string fullPath = Path.GetFullPath(Path.Combine(curDir, entry.FullName));
 
-                web.DownloadFile(downloadUrl, file.FullName);
-                using var zipFile = ZipFile.OpenRead(file.FullName);
-                foreach (ZipArchiveEntry entry in zipFile.Entries) {
-                    string fullPath = Path.GetFullPath(Path.Combine(curDir, entry.FullName));
-
-                    if (Path.GetFileName(fullPath).Length == 0) {
-                        Directory.CreateDirectory(fullPath);
-                    } else {
-                        Directory.CreateDirectory(Path.GetDirectoryName(fullPath));
-                        entry.ExtractToFile(fullPath, overwrite: true);
+                        if (Path.GetFileName(fullPath).Length == 0) {
+                            Directory.CreateDirectory(fullPath);
+                        } else {
+                            Directory.CreateDirectory(Path.GetDirectoryName(fullPath));
+                            entry.ExtractToFile(fullPath, overwrite: true);
+                        }
                     }
+                    zipFile.Dispose();
+
+                    file.Delete();
+
+                    logger.Log($"Successfully updated mod to version {version}!");
+                    return true;
+                } else {
+                    logger.Log($"Detected remote version {version} is not newer than local version {curVersion}");
                 }
-                zipFile.Dispose();
-
-                file.Delete();
-
-                logger.Log($"Successfully updated mod to version {version}!");
-                return true;
             } catch (Exception ex) {
                 logger.Log($"Error trying to update mod: \n{ex.ToString()}");
             }
